frontend_app/requirements.txt
Pillow==9.5.0
opencv-python
numpy
pyserial